validation_ch7_v2/scripts/niveau4_rl_performance/requirements.txt
# Section 7.6 RL Performance - Clean Architecture Dependencies

# RL Framework
stable-baselines3==2.3.2
gymnasium==0.29.1

# Core scientific computing
numpy>=1.24.0

# Configuration
pyyaml==6.0.1

# Structured Logging (Innovation 7)
structlog==24.1.0

# CLI
click==8.1.7

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Optional: Advanced monitoring
# tensorboard==2.16.2  # Pour visualisation entraînement RL
# matplotlib>=3.7.0  # Pour visualisations graphiques
